+++ RELEASE_NOTES.md
@@ -24,6 +24,7 @@ Due to a bug found in the new autocomplete provider, the old `StandardAutocomple
 
 * `KeyboardAppearance` has a new `keyboardBackground` property.
 * `KeyboardBackgroundStyle` is a new style that can be used to the background of a keyboard.
+* `KeyboardButtonStyle` properties are now optional to allow for overrides.
 
 ### 👑 Pro changes